sparrow-ml/lemming/requirements.txt
llama-index==0.9.33
langchain==0.1.1
sentence-transformers
weaviate-client
pypdf
python-box
unstructured[all-docs]
typer[all]
fastapi==0.108.0
uvicorn[standard]
# add pdf2htmlEX support for RAG pipeline with LlamaIndex and Unstructured